driver core: Move the "authorized" attribute from USB/Thunderbolt to core
Currently bus drivers like "USB" or "Thunderbolt" implement a custom
version of device authorization to selectively authorize the driver
probes. Similar requirements have been discussed in the PCI [1]
community for PCI bus drivers as well. So moved the "authorized"
attribute support to the driver core in order to allow it to be
used by other subsystems / buses.

There should be no user-visible change in the location or semantics of
attributes for USB/Thunderbolt devices.

[1]: https://lore.kernel.org/r/20210804161839.3492053-1-Jonathan.Cameron@huawei.com

Signed-off-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>