mm: Add prot_guest_authorized() interface
Protected Guest platform has a requirement to allow only a trusted list
of devices. So add an interface to allow the bus driver check with arch
code to determine the default status of the authorized device.

If ARCH code implements prot_guest_authorized(), it will return true or
false based on arch specific device allow or deny list.

This is a preparatory patch for adding x86 ARCH specific
prot_guest_authorized() support.

Signed-off-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>